requirements.txt
Flask==2.3.2
pika==1.3.2
google-api-python-client==2.77.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
python-dateutil
mysql-connector-python
pytest
python-dotenv>=0.21.0